arena utils test

+++ include/sealloc/arena.h
@@ -39,7 +39,7 @@ void arena_init(arena_t *arena);
 chunk_t *arena_allocate_chunk(arena_t *arena);
 void arena_deallocate_chunk(arena_t *arena, chunk_t *chunk);
 chunk_t *arena_get_chunk_from_ptr(arena_t *arena, void *ptr);
-bin_t *arena_get_bin_by_reg_size(arena_t *arena, unsigned reg_size);
+bin_t *arena_get_bin_by_reg_size(arena_t *arena, uint16_t reg_size);
 huge_chunk_t *arena_find_huge_mapping(arena_t *arena, void *huge_map);
 huge_chunk_t *arena_allocate_huge_mapping(arena_t *arena, size_t size);
 void arena_deallocate_huge_mapping(arena_t *arena, void *huge_map);
